move _url functions to network.py
